Replace old Flywheel class
