+++ src/main/java/frc/robot/subsystems/shooter/FlyWheel.java
@@ -1,43 +1,59 @@
 package frc.robot.subsystems.shooter;
 
-import edu.wpi.first.math.MathUtil;
-import edu.wpi.first.math.controller.PIDController;
-import edu.wpi.first.math.controller.SimpleMotorFeedforward;
-
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-
-import frc.robot.Constants;
+import java.util.function.Supplier;
 
 import com.ctre.phoenix6.controls.VoltageOut;
 import com.ctre.phoenix6.hardware.TalonFX;
 
-public class FlyWheel extends SubsystemBase {
-    PIDController m_RightPIDController = new PIDController(ShooterConstants.RIGHT_FLYWHEEL_P, 0, 0);
-    PIDController m_LeftPIDController = new PIDController(ShooterConstants.LEFT_FLYWHEEL_P, 0, 0);
+import edu.wpi.first.math.MathUtil;
+import edu.wpi.first.math.controller.PIDController;
+import edu.wpi.first.math.controller.SimpleMotorFeedforward;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.wpilibj.DriverStation;
+import edu.wpi.first.wpilibj.DriverStation.Alliance;
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.robot.Constants;
+import frc.robot.StateManager.RobotStateMachine;
+import frc.robot.utils.Utils;
+
+public class FlyWheel2 extends SubsystemBase{
+
+    PIDController m_RightPIDController =
+        new PIDController(ShooterConstants.RIGHT_FLYWHEEL_P, 0, 0);
+    PIDController m_LeftPIDController =
+        new PIDController(ShooterConstants.LEFT_FLYWHEEL_P, 0, 0);
     
-    SimpleMotorFeedforward m_RightFeedforwardController = new SimpleMotorFeedforward(0, 0.00174, 0);
-    SimpleMotorFeedforward m_LeftFeedforwardController = new SimpleMotorFeedforward(0, 0.00173, 0);
+    SimpleMotorFeedforward m_RightFeedforwardController =
+        new SimpleMotorFeedforward(0, 0.00174, 0);
+    SimpleMotorFeedforward m_LeftFeedforwardController =
+        new SimpleMotorFeedforward(0, 0.00173, 0);
+
     private final TalonFX m_RightFlywheelMotor;
     private final TalonFX m_LeftFlywheelMotor;
 
     private VoltageOut m_rightRequest = new VoltageOut(0.0);
+    private VoltageOut m_leftRequest = new VoltageOut(0.0);
 
-    boolean intaking = false;
-    boolean readyToShoot = false;
-    boolean isRunning = false;
-    boolean stopped = true;
+    private boolean activated = false;
+    private boolean shooting = false;
 
-    public FlyWheel() {;
+    private Supplier<RobotStateMachine> stateMachineSupplier;
+    private Supplier<Pose2d> poseSupplier;
+
+    public FlyWheel2(Supplier<RobotStateMachine> stateMachineSupplier, Supplier<Pose2d> poseSupplier) {
+        this.stateMachineSupplier = stateMachineSupplier;
+        this.poseSupplier = poseSupplier;
+
         m_RightFlywheelMotor = new TalonFX(Constants.FlyWheel.SHOOTER_RIGHT_ID);
         m_LeftFlywheelMotor = new TalonFX(Constants.FlyWheel.SHOOTER_LEFT_ID);
 
         m_RightFlywheelMotor.setInverted(false);
         m_LeftFlywheelMotor.setInverted(true);
+
         m_RightPIDController.setTolerance(200);
         m_LeftPIDController.setTolerance(200);
     }
-    
+
     public double getRightFlywheelRPM() {
         return (m_RightFlywheelMotor.getVelocity().getValueAsDouble() * 60) / 0.8;
     }
@@ -46,144 +62,66 @@ public class FlyWheel extends SubsystemBase {
         return (m_LeftFlywheelMotor.getVelocity().getValueAsDouble() * 60) / 0.8;
     }
 
-    public void intake() {
-        intaking = true;
-        isRunning = false;
-        stopped = false;
-    }
-    public void run() {
-        isRunning = true;
-        stopped = false;
-        intaking = false;
+    public void start() {
+        activated = true;
     }
 
-    private void updateValues() {
+    public void shoot() {
+        shooting = true;
+        activated = true;
 
-        double rightVoltage = m_RightPIDController.calculate(
-                    getRightFlywheelRPM(),
-                    6500
-                );
-
-        double leftVoltage = m_LeftPIDController.calculate(
-                getLeftFlywheelRPM(), 
-                6500
-            );
-        
-        m_RightFlywheelMotor.setControl(
-            m_rightRequest.withOutput(
-                rightVoltage + m_RightFeedforwardController.calculate(6500)
-            )
-        );
-
-        m_LeftFlywheelMotor.setVoltage(
-            leftVoltage + m_LeftFeedforwardController.calculate(6500)
-        );
-
-        SmartDashboard.putNumber("Right PID Output", rightVoltage);
-        SmartDashboard.putNumber("Left PID Output", rightVoltage);
     }
-
-    private void updateIdle() {
-        double rightVoltage = m_RightPIDController.calculate(
-                    getRightFlywheelRPM(),
-                    0
-                );
-
-        double leftVoltage =  m_LeftPIDController.calculate(
-                getLeftFlywheelRPM(), 
-                0
-            );
-        
-        m_RightFlywheelMotor.setControl(
-            m_rightRequest.withOutput(
-                MathUtil.clamp(
-                    rightVoltage + m_RightFeedforwardController.calculate(0),
-                    0,
-                    12
-                )
-            )
-        );
-        m_LeftFlywheelMotor.setVoltage(
-            MathUtil.clamp(
-            leftVoltage + m_LeftFeedforwardController.calculate(0),
-            0,
-            12
-            )
-        );
-    }
-
-        private void updateIntake() {
-        double rightVoltage = m_RightPIDController.calculate(
-                    getRightFlywheelRPM(),
-                    -2000
-                );
-
-        double leftVoltage =  m_LeftPIDController.calculate(
-                getLeftFlywheelRPM(), 
-                -1500
-            );
-        
-        m_RightFlywheelMotor.setControl(
-            m_rightRequest.withOutput(
-                MathUtil.clamp(
-                    rightVoltage + m_RightFeedforwardController.calculate(-1000),
-                    -12,
-                    0
-                )
-            )
-        );
-        m_LeftFlywheelMotor.setVoltage(
-            MathUtil.clamp(
-            leftVoltage + m_LeftFeedforwardController.calculate(-1000),
-            -12,
-            0
-            )
-        );
-    }
-
-    public void idle() {
-        intaking = false;
-        isRunning = false;
-        stopped = false;
-    }
-
     public void stop() {
-        intaking = false;
-        stopped = true;
-        isRunning = false;
+        activated = false;
     }
 
-    public boolean isReady() {
-        return m_LeftPIDController.atSetpoint() && m_RightPIDController.atSetpoint();
+    private void setFlywheelSpeeds(double leftSpeedRequest, double rightSpeedRequest) {
+        m_RightFlywheelMotor.setControl(m_rightRequest.withOutput(
+            m_RightPIDController.calculate(getRightFlywheelRPM(), rightSpeedRequest) +
+                m_RightFeedforwardController.calculate(rightSpeedRequest)
+        ));
+
+        m_LeftFlywheelMotor.setControl(m_leftRequest.withOutput(
+            m_LeftPIDController.calculate(getRightFlywheelRPM(), leftSpeedRequest) +
+                m_LeftFeedforwardController.calculate(leftSpeedRequest)
+        ));
     }
 
     @Override
     public void periodic() {
-        
-        if (stopped) {
+
+        if (!activated) {
             m_RightFlywheelMotor.stopMotor();
             m_LeftFlywheelMotor.stopMotor();
-        } else if (isRunning) {
-            updateValues();
-        } else if (intaking) {
-            updateIntake();
-        } else {
-            updateIdle();
+            return;
         }
 
-        SmartDashboard.putNumber("Right Flywheel rpm", getRightFlywheelRPM());
-        SmartDashboard.putNumber("Left Flywheel rpm", getLeftFlywheelRPM());
+        switch (stateMachineSupplier.get()) {
+            case NORMAL:
 
-        SmartDashboard.putNumber("Right Flywheel PID error", m_RightPIDController.getPositionError());
-        SmartDashboard.putNumber("Left flywheel PID error", m_LeftPIDController.getPositionError());
-        SmartDashboard.putNumber("Right flywheel setpoint", m_RightPIDController.getSetpoint());
-        SmartDashboard.putNumber("Left Flywheel setpoint", m_LeftPIDController.getSetpoint());
-        if (isReady()) {
-            readyToShoot = true;
-        } else {
-            readyToShoot = false;
+                if (Utils.getDistance(poseSupplier) <= 5) {
+                    setFlywheelSpeeds(4000, 4000);
+                    break;
+                }
+                
+                if (!shooting) {
+                    m_LeftFlywheelMotor.stopMotor();
+                    m_RightFlywheelMotor.stopMotor();
+                    break;
+                }
+
+                setFlywheelSpeeds(6500, 6500);
+                break;
+            case SHUTTLE:
+                if (Math.abs(Utils.getDistanceFromOppWing(poseSupplier, DriverStation.getAlliance().get())) <= 2) { //TODO: Don't call get alliance every time
+                    //TODO: Linear Regression
+                }
+                break;
+            case A10BRRRRR:
+                setFlywheelSpeeds(3000, 3000);
+                break;
+            case AMP: //The shooter should never ramp up when AMPing
+                break;
         }
-
-        SmartDashboard.putBoolean("Ready to shoot", readyToShoot);
     }
 }
